upgrade sdk change implementation to compare image (send image before after from ar page) change to binding to easier clear view (prevent leaking when do fragment transaction)

+++ features/merchant/product_ar/src/main/java/com/tokopedia/product_ar/util/ProductArMapper.kt
@@ -156,26 +156,26 @@ object ProductArMapper {
                 data.colorB.toIntSafely())
 
         val product = MFEMakeupProduct().also {
-            it.gloss = data.getGlossFormula().toInt()
-            it.glossDetail = data.getGlossDetailFormula().toInt()
-            it.wetness = data.getWetnessFormula().toInt()
-            it.glitter = data.getSparkleAFormula().toInt()
+            it.gloss = data.getGlossFormula()
+            it.glossDetail = data.getGlossDetailFormula()
+            it.wetness = data.getWetnessFormula()
+            it.glitter = data.getSparkleAFormula()
             it.glitterColor = data.getGlitterColor()
             it.color = color
             it.matteness = data.matteness
-            it.glitterDensity = data.getGlitterDensityFormula().toInt()
+            it.glitterDensity = data.getGlitterDensityFormula()
             it.glitterSize = data.getGlitterSizeFormula()
-            it.glitterColorVariation = data.getGlitterColorVariationFormula().toInt()
-            it.glitterSizeVariation = data.getGlitterSizeVariationFormula().toInt()
-            it.glitterBaseReflectivity = data.getGlitterBaseReflectivityFormula().toInt()
-            it.envMappingIntensity = data.getEnvMappingtFormula().toInt()
+            it.glitterColorVariation = data.getGlitterColorVariationFormula()
+            it.glitterSizeVariation = data.getGlitterSizeVariationFormula()
+            it.glitterBaseReflectivity = data.getGlitterBaseReflectivityFormula()
+            it.envMappingIntensity = data.getEnvMappingtFormula()
             it.envMappingColor = data.getEnvMappingColor()
             it.envMappingBumpIntensity = data.getEnvBumpDensityFormula()
             it.envMappingCurve = data.getEnvMappingCurveFormula()
             it.envMappingRotationY = data.envMappingRotationY
             it.metallicIntensity = data.metallicIntensity
             it.vinylIntensity = data.vinylIntensity
-            it.amount = data.getAmountFormula().toInt()
+            it.amount = data.getAmountFormula()
         }
 
         MFEMakeupLook().apply {
